refactor: move apply default prop logic to renderer
