+++ packages/core/src/index.ts
@@ -4,4 +4,3 @@ export * from './prompt-generator';
 export * from './delta-json-path-selector';
 export * from './stream-pattern-extractor';
 export * from './repair-json';
-export * from './material-defaults';
