--- packages/core/src/material-defaults/index.ts
@@ -1,1 +1,0 @@
-export * from './material-defaults';
