+++ packages/core/src/protocols/index.ts
@@ -2,3 +2,4 @@ export * from './materials';
 export * from './schema';
 export * from './render-config';
 export * from './chat';
+export * from './default-props-map';
