+++ packages/frameworks/vue/src/renderer/index.ts
@@ -1,4 +1,3 @@
 export { default as GenuiRenderer } from './SchemaCardRenderer.vue'; 
 export * from './config.js'
-export * from './material-defaults.js';
 export * from './renderer.types.js';
